libs/python/pyproject.toml
[project]
name = "book-creator-schemas"
version = "0.1.0"
description = "Shared Pydantic models and utilities for the book creator platform"
authors = [{name = "Book Creator Team"}]
requires-python = ">=3.11"
dependencies = [
    "pydantic>=2.8,<3",
    "pydantic[email]>=2.8,<3",
    "httpx>=0.27,<0.28",
    "prometheus-client>=0.20,<0.21",
]

[project.optional-dependencies]
providers = [
    "openai>=1.51.0,<2",
    "google-genai>=1.1.0,<2",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.pydantic-mypy]
init_forbid_extra = true
warn_required_dynamic_aliases = true
warn_untyped_fields = true

[tool.setuptools.packages.find]
include = [
    "book_creator_schemas",
    "book_creator_schemas.*",
    "book_creator_providers",
    "book_creator_providers.*",
    "book_creator_observability",
    "book_creator_observability.*",
]
